refactor(token): extract allowance logic into allowance.rs module (closes #539)
Centralises get_allowance, set_allowance, and deduct_allowance in a new
allowance.rs module. deduct_allowance preserves the original expiration_ledger
so the entry TTL is not inadvertently reset on every spend. All callers in
token_interface.rs (approve, transfer_from, burn_from) and lib.rs (allowance
query) updated to use the module.
